+++ settings.js
@@ -15,6 +15,7 @@
         caseBulkBuy: true,
         caseAutoOpen: true,
         tradeSearch: true,
+        tradeFloatOverlays: true,
         skinLock: true,
     };